@@ -1,6 +1,6 @@
 module github.com/sarpt/mpv-web-api
 
-go 1.18
+go 1.21
 
 require (
 	github.com/fsnotify/fsnotify v1.5.1
